临时修改

+++ src/main/java/com/dianying/client/api/AppApi.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.annotation.Resource;
 import java.io.File;
 import java.io.IOException;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -41,9 +42,15 @@ public class AppApi {
         Long dataId = null;
         try {
             bean = JSONObject.parseObject(data,AppBean.class);
-            bean.setLogTime(System.currentTimeMillis()/1000);
-            int count = appDao.exist(bean.getAppcode(),bean.getType());
-            if (count<=0) {
+            if (bean.getStartx()<720&&bean.getStarty()<1280&&bean.getWidth()<100&&bean.getHeight()<100) {
+                bean.setLogTime(System.currentTimeMillis() / 1000);
+                Integer type = appDao.getType(bean.getAppcode());
+                if (type == null) {
+                    bean.setType(0);
+                } else {
+                    bean.setType(type + 1);
+                }
+                bean.setDir(bean.getAppcode() + "_" + bean.getType());
                 dataId = appDao.save(bean);
                 if (dataId == null) {
                     errorMsg = "保存失败";
@@ -51,7 +58,7 @@ public class AppApi {
                     errorMsg = "当前类型已经存在";
                 }
             }else {
-                errorMsg = "当前类型已经存在";
+                errorMsg = "起点或者长宽不符合要求";
             }
         }catch (Exception e){
             errorMsg = "提交的信息数据格式异常";
@@ -70,14 +77,18 @@ public class AppApi {
         AppBean bean = null;
         try {
             bean = JSONObject.parseObject(data, AppBean.class);
-            bean.setLogTime(System.currentTimeMillis()/1000);
-            int count = appDao.exist(bean.getAppcode(),bean.getType());
-            if(count<=0) {
-                if (!appDao.update(bean)) {
-                    errorMsg = "修改数据";
+            if (bean.getStartx()<720&&bean.getStarty()<1280&&bean.getWidth()<100&&bean.getHeight()<100) {
+                bean.setLogTime(System.currentTimeMillis() / 1000);
+                int count = appDao.exist(bean.getAppcode(), bean.getType());
+                if (count > 0) {
+                    if (!appDao.update(bean)) {
+                        errorMsg = "修改数据失败";
+                    }
+                } else {
+                    errorMsg = "该类型不存在";
                 }
             }else {
-                errorMsg = "该类型不存在";
+                errorMsg = "起点或者长宽不符合要求";
             }
         }catch (Exception e){
             errorMsg = "提交的信息数据格式异常";
@@ -102,6 +113,14 @@ public class AppApi {
             page = Page.parseFromJson(data);
         }
         List<Map<String,Object>> list = appDao.getAllApp(page);
+        Iterator<Map<String, Object>> it = list.iterator();
+        while (it.hasNext()){
+            Map<String,Object> item = it.next();
+            String dir = (String) item.get("dir");
+            File file = new File(Config.UPLOAD_DIR,dir);
+            System.out.println(file.length());
+            item.put("imageNum",(file.list() == null ? 0 : (file.list()).length));
+        }
         return APIUtil.genPageDataObject(list,page);
     }
 
@@ -114,12 +133,21 @@ public class AppApi {
     @RequestMapping("/delete")
     public Object delete(@RequestParam(value = "appcode",required = true) String appcode,
                          @RequestParam(value = "type",required = true) String type){
+        String errorMsg = StringUtils.EMPTY;
         AppBean bean = appDao.queryforBean(appcode,type);
-        FileService.delete(bean.getDir());
-       if(appDao.delete(appcode,type)){
-            return APIUtil.genSuccessResult();
+        if (bean != null){
+            if(appDao.delete(appcode,type)){
+                FileService.delete(bean.getDir());
+            }else {
+                errorMsg = "删除失败";
+            }
         }else {
-            return APIUtil.genErrorResult("sql执行失败");
+            errorMsg = "类型不存在";
+        }
+        if (StringUtils.isEmpty(errorMsg)) {
+            return APIUtil.genSuccessResult();
+        } else {
+            return APIUtil.genErrorResult(errorMsg);
         }
     }
 
@@ -131,24 +159,24 @@ public class AppApi {
      */
     @RequestMapping("/create")
     public Object creatIconFile(String appcode, String type) {
-        String msg = StringUtils.EMPTY;
+        String errorMsg = StringUtils.EMPTY;
         AppBean bean = appDao.queryforBean(appcode, type);
         String iconFile = "";
         if (bean == null) {
-            msg = "记录不存在";
-            return APIUtil.genErrorResult(msg);
+            errorMsg = "记录不存在";
+            return APIUtil.genErrorResult(errorMsg);
         }
         iconFile = bean.getAppcode() + "-" + bean.getType();
         File dirFile = new File(Config.UPLOAD_DIR + "/" + bean.getDir());
         if (!dirFile.exists()) {
-            msg = dirFile.getAbsolutePath()+"文件夹不存在";
-            return APIUtil.genErrorResult(msg);
+            errorMsg = dirFile.getAbsolutePath()+"文件夹不存在";
+            return APIUtil.genErrorResult(errorMsg);
 
         }
         File[] files = dirFile.listFiles();
         if (files == null || files.length == 0) {
-            msg = bean.getDir() + "中没有图片";
-            return APIUtil.genErrorResult(msg);
+            errorMsg = bean.getDir() + "中没有图片";
+            return APIUtil.genErrorResult(errorMsg);
         }
 
         long[] arr = new long[files.length];
@@ -168,12 +196,12 @@ public class AppApi {
         }
         boolean result = IconUtil.saveCharacter(desWidth, desHeight, startX, startY, iconFile, arr);
         if (!result) {
-            msg = "生成特征文件失败";
+            errorMsg = "生成特征文件失败";
         }
-        if (StringUtils.isEmpty(msg)) {
+        if (StringUtils.isEmpty(errorMsg)) {
             return APIUtil.genDataResult(iconFile);
         } else {
-            return APIUtil.genErrorResult(msg);
+            return APIUtil.genErrorResult(errorMsg);
         }
     }
 
@@ -185,8 +213,5 @@ public class AppApi {
         }catch (Exception e){
             return APIUtil.genErrorResult(e.toString());
         }
-
     }
-
-
 }
